requirement.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.5.2
bcrypt==5.0.0
certifi==2025.10.5
click==8.1.8
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.11
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.44.0
typer==0.20.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
